Optimized collision detection logic.
Using dmd -profile -debug, the run time of isLocationValidTile was
dropped from 8090 to 3550.  Instead of checking the entire image
boundary, the logic was made to test only the new area to be occupied,
try a greedy move, and back off if there's a collision.

Also, several functions which used struct parameters were changed to
use 'in', which shaved off another 84 time units.