Got demo GtkD window to compile and run.
This is the humble beginning of the level editor.
Major flaws were discovered in DSSS, including a total
lack of support for D2 'version'.  The D2 keyword 'D_Version2'
is overriden and it prevents GtkD from compiling correctly.

Chances are that building may have to revert to a combination
of make and rdmd.